SNE-72 Shifted banner and skipnav into the header.js so that the order was easier to control, and so that they all ended up in the <header>
